Update bannergrab.py into python 3
